Makefile: Duplicate the theme into the elm folders.
Instead of only installing it into the e theme folders we now duplicate
the theme into elms theme folders as well. This allows elementary_config
to choose it as a theme without manually copying files around.

In the future we aim for sharing one theme for both e and elm. Once we have
reached this point we can remove this.